forest_cover_prediction/requirements.txt
scikit-learn>=1.0.0
xgboost>=1.5.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.0.0
joblib>=1.0.0